+++ plugin/src/adapter50/kotlin/org/jlleitschuh/gradle/ktlint/worker/KtLintInvocation50.kt
@@ -2,9 +2,11 @@ package org.jlleitschuh.gradle.ktlint.worker
 
 import com.pinterest.ktlint.cli.ruleset.core.api.RuleSetProviderV3
 import com.pinterest.ktlint.rule.engine.api.Code
+import com.pinterest.ktlint.rule.engine.api.EditorConfigOverride
 import com.pinterest.ktlint.rule.engine.api.KtLintRuleEngine
 import com.pinterest.ktlint.rule.engine.api.LintError
 import com.pinterest.ktlint.rule.engine.core.api.RuleProvider
+import com.pinterest.ktlint.rule.engine.core.api.editorconfig.EditorConfigProperty
 import java.io.File
 import java.util.ServiceLoader
 
@@ -12,10 +14,22 @@ class KtLintInvocation50(
     private val engine: KtLintRuleEngine
 ) : KtLintInvocation {
     companion object Factory : KtLintInvocationFactory {
-        fun initialize(): KtLintInvocation {
-            val engine = KtLintRuleEngine(
-                ruleProviders = loadRuleSetsFromClasspathWithRuleSetProviderV3()
-            )
+        fun initialize(editorConfigOverrides: Map<String, String>): KtLintInvocation {
+            val ruleProviders = loadRuleSetsFromClasspathWithRuleSetProviderV3()
+            val engine = if (editorConfigOverrides.isEmpty()) {
+                KtLintRuleEngine(ruleProviders = ruleProviders)
+            } else {
+                KtLintRuleEngine(
+                    ruleProviders = ruleProviders,
+                    editorConfigOverride = EditorConfigOverride.from(
+                        *editorConfigOverrides
+                            .mapKeys { ruleProviders.findEditorConfigProperty(it.key) }
+                            .entries
+                            .map { it.key to it.value }
+                            .toTypedArray()
+                    )
+                )
+            }
             return KtLintInvocation50(engine)
         }
 
@@ -51,4 +65,22 @@ class KtLintInvocation50(
 
 internal fun LintError.toSerializable(): SerializableLintError {
     return SerializableLintError(line, col, ruleId.value, detail, canBeAutoCorrected)
+}
+
+private fun Set<RuleProvider>.findEditorConfigProperty(propertyName: String): EditorConfigProperty<*> {
+    val properties =
+        map { it.createNewRuleInstance() }
+            .flatMap { it.usesEditorConfigProperties }
+            .distinct()
+    return properties
+        .find { it.type.name == propertyName }
+        ?: throw RuntimeException(
+            properties
+                .map { it.type.name }
+                .sorted()
+                .joinToString(
+                    prefix = "Property with name '$propertyName' is not found in any of given rules. Available properties:\n\t",
+                    separator = "\n\t"
+                ) { "- $it" }
+        )
 }